+++ HelloVectors/HelloVectors_win32_cpp/HelloVectors_win32_cpp.vcxproj
@@ -154,20 +154,19 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClInclude Include="desktopwindow.h" />
-    <ClInclude Include="HelloVectors_win32.h" />
-    <ClInclude Include="ICompositionSource.h" />
+    <ClInclude Include="LottieTool\ICompositionSource.h" />
+    <ClInclude Include="LottieTool\SquareCircleMorph.h" />
     <ClInclude Include="pch.h" />
-    <ClInclude Include="SquareCircleMorph.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="HelloVectors_win32.cpp" />
+    <ClCompile Include="LottieTool\SquareCircleMorph.cpp" />
     <ClCompile Include="pch.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">Create</PrecompiledHeader>
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Release|x64'">Create</PrecompiledHeader>
     </ClCompile>
-    <ClCompile Include="SquareCircleMorph.cpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
