+++ HelloVectors/HelloVectors_win32_cpp/LottieTool/ICompositionSource.h
